refactor(threads): useThreadMessaging 主干外移 memory pick 闸门段（runMemoryPickGate）
